split unbinned / binned cluster likelihoods
