+++ soliket/clusters/__init__.py
@@ -1,1 +1,1 @@
-from .clusters_both import UnbinnedClusterLikelihood #, BinnedClusterLikelihood  # noqa: F401
+from .clusters import BinnedClusterLikelihood, UnbinnedClusterLikelihood  # noqa: F401
