+++ soliket/tests/data/toy_cashc.txt
@@ -1,0 +1,4 @@
+0.		100		2
+0.2		100		4
+0.4		100		5
+0.6		100		2
